+++ src/main/java/com/dadasoft/gestorDeGastos/entity/catalogo/TipoDeMovimientoDAO.java
@@ -1,8 +1,6 @@
 package com.dadasoft.gestorDeGastos.entity.catalogo;
 
-import lombok.AllArgsConstructor;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import javax.persistence.Entity;
