+++ src/main/java/com/dadasoft/gestorDeGastos/service/impl/CategoriaServiceImpl.java
@@ -1,7 +1,8 @@
 package com.dadasoft.gestorDeGastos.service.impl;
 
 import com.dadasoft.gestorDeGastos.api.CategoriaApi;
-import com.dadasoft.gestorDeGastos.entity.CategoriaDAO;
+import com.dadasoft.gestorDeGastos.entity.catalogo.CategoriaDAO;
+import com.dadasoft.gestorDeGastos.exception.CategoriaException;
 import com.dadasoft.gestorDeGastos.repository.ICategoriaRepo;
 import com.dadasoft.gestorDeGastos.service.ICategoriaService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -9,6 +10,7 @@ import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class CategoriaServiceImpl implements ICategoriaService {
@@ -39,6 +41,20 @@ public class CategoriaServiceImpl implements ICategoriaService {
 		List<CategoriaApi> listCategoriesApi;
 		listCategoriesApi = convert(categoriaRepo.findAll());
 		return listCategoriesApi;
+	}
+
+	public CategoriaApi deshabilitarCategoria(String id) throws CategoriaException {
+		CategoriaApi categoriaApi;
+		// Optional<CategoriaDAO> opCategoriaDAO = Optional.ofNullable(categoriaRepo.findById(Long.parseLong(id))).orElseThrow(() -> new CategoriaException(CategoriaException.CATEGORIA_NOT_FOUND_CODE,
+		// 		CategoriaException.CATEGORIA_NOT_FOUND_MSG + " (" + id + ")"));
+
+		Optional<CategoriaDAO> opCategoriaDA = categoriaRepo.findById(Long.parseLong(id));
+		CategoriaDAO categoriaDAO = opCategoriaDA.orElseThrow(() -> new CategoriaException(CategoriaException.CATEGORIA_NOT_FOUND_CODE,
+																							CategoriaException.CATEGORIA_NOT_FOUND_MSG + " (" + id + ")"));
+
+		categoriaDAO.setEnable(false);
+		categoriaApi = convert(categoriaRepo.saveAndFlush(categoriaDAO));
+		return categoriaApi;
 	}
